Fix #125 - Support natural key when working with interfaces
The NSoT API supports identifying most things by either their unique ID
number, or by a natural key; the natural key for interfaces being the
combination of the device's hostname and the interface's name. Example:

  * Device: switch-nyc3
  * Interface: Ethernet2
  * Natural Key: switch-nyc3:Ethernet2

Since dropbox/nsot#253 added caching of the device hostname on the
Interface object, lookups using that natural key are much easier.

Similar to how the API overloads the ID to be either one of those
things, I've mirrored that in the CLI. It would probably be more
user-friendly to let the user do something like specify a device
with `-D` and the interface name with `-n` or something, but that
causes a conflict with `update` since it uses `-n` to update the name of
the interface.

Given that conflict, that would have created a special case for `update`
and I felt like consistency across the various commands takes precedence.
Requiring the user to concatenate the device hostname and interface name
is a little annoying, but I tried to document that in a clear way.