Add fix for slashes in Circuit URLs
There was a problem in NSoT where a slash in a Circuit URL would cause
issues with the URL parser. The workaround was to create a `name_slug`
field on Circuit and use that as the natural key.

To support this in pynsot, every time the user supplies a circuit name
for the purposes of looking up a circuit, we need to slugify it in the
same way that NSoT does. This is why `nsot` moved from
`requirements-dev.txt` to `requirements.txt`, and the version was bumped
to 1.1.1 which added the `slugify` function.

More details can be found here:
https://github.com/dropbox/nsot/pull/259