Add support for Circuits (#129)
Now that we have circuits in NSoT, the CLI needs to support them. This handles all the CRUD operations, as well as subcommands for list in the API: devices, addresses, and interfaces.

I also did a bit of refactoring of the tests to support doing set of tests per file (e.g. circuits in their own file). Since I added a set of fixtures for the Circuit tests, those went into their own file. tests/fixtures.py was moved to test/fixtures/__init__.py to support this and still maintain backwards compatibility of imports in the existing tests.