Change requirements*txt to use Compatible Release specifiers
This is a similar change to https://github.com/dropbox/nsot/pull/265

For the `nsot` requirement, I left it at `~=1.1.0` since API
deprecations could happen if we jump too far ahead on minor versions. I
also added a minimum version of `nsot` to ensure that tests run against
a version that has the APIs we need.