Change requirements*txt to use Compatible Release specifiers (#131)
* Change requirements*txt to use Compatible Release specifiers

This is a similar change to https://github.com/dropbox/nsot/pull/265

For the `nsot` requirement, I left it at `~=1.1.0` since API
deprecations could happen if we jump too far ahead on minor versions. I
also added a minimum version of `nsot` to ensure that tests run against
a version that has the APIs we need.

* Add minimum version for `py` package

Argh, I hit this when making this change in `nsot` but missed it here:
https://github.com/dropbox/nsot/pull/265/commits/101e321860c662db3e28145f8f971e2acf8bba2a

Basically there's an implicit dependency somewhere on `py>=1.4.29` so we
need that in our `requirements.txt`.