Fix #119 - Add ability to use set queries on `list` subcommands (#133)
The use-case here is described by @jathanism in nsot#221, where single
objects should be able to be referenced via set query for the various
subcommands on `list`, such as `next_network` on `Network` objects.
Since these subcommands operate on a single object, we set `unique=True`
in the request so that the API will return an error if the set query
returns != 1 object.

The implementation for this is done in the `list` subcommand callback,
where we were already translating from a natural key into an ID to be
given to the subcommand request. The heavy lifting is done by a new
method on `pynsot.App`.

While I was in there, I refactored the old `set_query` method to have a
better name. If this is expected to create issues with people
programatically consuming pynsot, I can undo that refactor.